feat: career details page
