+++ src/page-building/career-details/career-details.stories.mdx
@@ -1,0 +1,47 @@
+import {
+  ArgsTable,
+  Canvas,
+  Meta,
+  Story
+} from '@storybook/addon-docs';
+import { JobDescription } from './job-description';
+import { CareerForm } from '../../components/career-form/career-form';
+
+<Meta title="Pages/Career Details" component={{ Image }} />
+
+# Career Details
+Here the parts to build the career details page.
+
+## Job Description
+We will receive multiple sections from the Personio API. Pass them to this
+component to render the job description.
+
+<Canvas withSource={'open'}>
+  <Story name="Job Description">
+    <JobDescription sections={[
+      {headline: "First", descriptionHtml: 'some <em>normal</em> html'},
+      {headline: "Second", descriptionHtml: `some <strong>other</strong> html`},
+      {headline: "Third", descriptionHtml: 'some <u>better</u> html'}
+    ]}/>
+  </Story>
+</Canvas>
+
+<ArgsTable of={JobDescription}/>
+
+
+## Career Form
+Pass in the required personio parameters to allow the form to submit
+applications to the personio application endpoint.
+
+<Canvas withSource={'open'}>
+  <Story name="Career Form">
+    <CareerForm
+      company_id="41230"
+      recruiting_channel_id="source-id"
+      access_token="some-access-token-for-personio"
+      job_position_id="personio-job-id"
+    />
+  </Story>
+</Canvas>
+
+<ArgsTable of={CareerForm}/>
